+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Fixed
 
+- 転送アクションのagent.urlとrecipient.urlがnullになる状態を解消
+
 ### Security
 
 ## v5.12.1 - 2022-04-06